+++ NEWS.md
@@ -1,5 +1,7 @@
 # dplyr (development version)
 
+* Empty `rowwise()` list-column elements now resolve to `logical()` rather than a random logical of length 1 (#7710).
+
 * `last_dplyr_warnings()` no longer prevents objects from being garbage collected (#7649).
 
 * Progress towards making dplyr conformant with the public C API of R (#7741).